Consolidate duplicate getDeepStateCopy tests
